shadows now cast correctly in the right directon, however it doesnt take into account whether the occlusion is in front or behind on the ray, its also quite noisy, some optimisation done.

+++ RayEngine/Circle_2D.hpp
@@ -23,7 +23,7 @@ public:
         return radius;
     }
 
-    sf::Vector2f getPosition() override {
+    sf::Vector2i getPosition() override {
         return position;
     }
 };
